+++ src/main/java/com/gestaooserp/dev/repository/ClienteRepository.java
@@ -3,5 +3,7 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteRepository  extends JpaRepository <Cliente,Integer> {
+import java.util.Optional;
+
+public interface ClienteRepository  extends JpaRepository <Cliente,Long> {
 }
