+++ src/main/java/com/gestaooserp/dev/repository/OrdemServicoRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.OrdemServico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrdemServicoRepository extends JpaRepository<OrdemServico,Integer> {
+public interface OrdemServicoRepository extends JpaRepository<OrdemServico,Long> {
 }
